resnet50

+++ mlssl_b06201018_v5/src/resnet50.py
@@ -443,7 +443,7 @@ class ResNet(nn.Module):
         # local_f is a list [local_f0, local_f1, ...]
         # local_offset = 8, grid_perside=[3,5] -> f[bs*8:bs*(8+18)], f[bs*(8+18):bs*(8+68)]
         cum_nofgrid = torch.cumsum(torch.Tensor([2*(nof_grid**2)
-                                                 for nof_grid in self.grid_per_side]), 0)
+                                                 for nof_grid in self.grid_per_side]), 0, dtype=int)
         local_f = [f[bs * (local_offset+cum_nofgrid[i-1]):bs * (local_offset + cum_nofgrid[i])]
                    if (i != 0) else f[bs * local_offset:bs * (local_offset + cum_nofgrid[i])]
                    for i in range(self.nmb_local_levels)]
